+++ picocli-codegen/src/main/java/picocli/codegen/docgen/manpage/ManPageGenerator.java
@@ -387,7 +387,7 @@ public class ManPageGenerator {
         List<ArgGroupSpec> groups = optionListGroups(spec);
         for (ArgGroupSpec group : groups) { options.removeAll(group.options()); }
 
-        if (spec.options().isEmpty()) {
+        if (options.isEmpty()) {
             return;
         }
         pw.printf("// tag::picocli-generated-man-section-options[]%n");